In Layer, define sa_oversize attribute.
This attribute defines the scanning area to be sa_oversize*max_bounding_box_dimension.
Currently, the scanning area side length (it's always a square) is simply equal to max_bounding_box_dimension.
So sa_oversize = 1.1 results in scanning area 10% larger than the max dimension of the bounding box containing all the shapes in the layer.